Refactoring, and adding logic to replace placeholders in default messages
